chore: move common logic
