+++ src/core_modules/capture-core/components/DataEntries/TeiRegistrationEntry/hooks/useBuildTeiPayload.js
@@ -3,13 +3,12 @@ import { useSelector } from 'react-redux';
 import { useMetadataForRegistrationForm } from '../../common/TEIAndEnrollment/useMetadataForRegistrationForm';
 import type { RenderFoundation } from '../../../../metaData';
 import { convertClientToServer, convertFormToClient } from '../../../../converters';
-import { capitalizeFirstLetter } from '../../../../../capture-core-utils/string';
 import { generateUID } from '../../../../utils/uid/generateUID';
 import { getDataEntryKey } from '../../../DataEntry/common/getDataEntryKey';
-import { FEATURETYPE } from '../../../../constants';
 import type {
     TeiPayload,
 } from '../../../Pages/common/TEIRelationshipsWidget/RegisterTei/DataEntry/TrackedEntityInstance/dataEntryTrackedEntityInstance.types';
+import { geometryType, getPossibleTetFeatureTypeKey, buildGeometryProp } from '../../common/TEIAndEnrollment/geometry';
 
 type Props = {
     trackedEntityTypeId: string,
@@ -21,35 +20,6 @@ type Props = {
 function getClientValuesForFormData(formValues: Object, formFoundation: RenderFoundation) {
     return formFoundation.convertValues(formValues, convertFormToClient);
 }
-
-function getPossibleTetFeatureTypeKey(serverValues: Object) {
-    return Object
-        .keys(serverValues)
-        .find(key => key.startsWith('FEATURETYPE_'));
-}
-
-function buildGeometryProp(key: string, serverValues: Object) {
-    if (!serverValues[key]) {
-        return undefined;
-    }
-    const type = capitalizeFirstLetter(key.replace('FEATURETYPE_', '').toLocaleLowerCase());
-    const coordinates = standardGeoJson(serverValues[key]);
-    return {
-        type,
-        coordinates,
-    };
-}
-
-const standardGeoJson = (geometry: Array<number> | { longitude: number, latitude: number }) => {
-    if (Array.isArray(geometry)) {
-        return geometry;
-    } else if (geometry.longitude && geometry.latitude) {
-        return [geometry.longitude, geometry.latitude];
-    }
-    return undefined;
-};
-
-const geometryType = formValuesKey => Object.values(FEATURETYPE).find(geometryKey => geometryKey === formValuesKey);
 
 const deriveAttributesFromFormValues = (formValues = {}) =>
     Object.keys(formValues)
