+++ src/core_modules/capture-core/components/DataEntries/common/TEIAndEnrollment/index.js
@@ -1,2 +1,3 @@
 // @flow
 export { getGeneratedUniqueValuesAsync, getUniqueValuesForAttributesWithoutValue } from './getGeneratedUniqueValuesAsync';
+export { geometryType, getPossibleTetFeatureTypeKey, buildGeometryProp } from './geometry';
